Renames FirebaseInit, removes FirebaseUI CDN from IndexHTML template
