+++ client/firebaseInit.js
